chore: setup muxing to allow migrating to plugin-framework
