+++ main.go
@@ -1,30 +1,47 @@
 package main // import "github.com/UpCloudLtd/terraform-provider-upcloud"
 
 import (
+	"context"
 	"flag"
+	"log"
 
 	"github.com/UpCloudLtd/terraform-provider-upcloud/upcloud"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
+	"github.com/hashicorp/terraform-plugin-framework/providerserver"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov5/tf5server"
+	"github.com/hashicorp/terraform-plugin-mux/tf5muxserver"
 )
 
 func main() {
-	var debugMode bool
-	var debugProviderAddr string
+	ctx := context.Background()
 
-	flag.BoolVar(&debugMode, "debug", false, "set to true to run the provider with support for debuggers like delve")
-	flag.StringVar(&debugProviderAddr, "debug-provider-addr", "registry.terraform.io/upcloudltd/upcloud",
-		"use same provider address as used in your configs")
+	var debug bool
 
+	flag.BoolVar(&debug, "debug", false, "set to true to run the provider with support for debuggers like delve")
 	flag.Parse()
 
-	opts := &plugin.ServeOpts{
-		ProviderFunc: func() *schema.Provider {
-			return upcloud.Provider()
-		},
-		Debug:        debugMode,
-		ProviderAddr: debugProviderAddr,
+	providers := []func() tfprotov5.ProviderServer{
+		providerserver.NewProtocol5(upcloud.New()),
+		upcloud.Provider().GRPCProvider,
 	}
 
-	plugin.Serve(opts)
+	muxServer, err := tf5muxserver.NewMuxServer(ctx, providers...)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	var serveOpts []tf5server.ServeOpt
+
+	if debug {
+		serveOpts = append(serveOpts, tf5server.WithManagedDebug())
+	}
+
+	err = tf5server.Serve(
+		"registry.terraform.io/upcloudltd/upcloud",
+		muxServer.ProviderServer,
+		serveOpts...,
+	)
+	if err != nil {
+		log.Fatal(err)
+	}
 }
