+++ upcloud/provider.go
@@ -2,176 +2,124 @@ package upcloud
 
 import (
 	"context"
-	"fmt"
-	"net/http"
+	"os"
 	"time"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
-
-	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/client"
-	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/service"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-
 	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/config"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/cloud"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/database"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/firewall"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/gateway"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/ip"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/kubernetes"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/loadbalancer"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/managedobjectstorage"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/network"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/objectstorage"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/router"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/server"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/servergroup"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/storage"
-	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/service/tag"
-
 	retryablehttp "github.com/hashicorp/go-retryablehttp"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/provider"
+	"github.com/hashicorp/terraform-plugin-framework/provider/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
-func Provider() *schema.Provider {
-	return &schema.Provider{
-		Schema: map[string]*schema.Schema{
-			"username": {
-				Type:        schema.TypeString,
+const (
+	usernameDescription       = "UpCloud username with API access. Can also be configured using the `UPCLOUD_USERNAME` environment variable."
+	passwordDescription       = "Password for UpCloud API user. Can also be configured using the `UPCLOUD_PASSWORD` environment variable."
+	requestTimeoutDescription = "The duration (in seconds) that the provider waits for an HTTP request towards UpCloud API to complete. Defaults to 120 seconds"
+)
+
+type upcloudProviderModel struct {
+	Username          types.String `tfsdk:"username"`
+	Password          types.String `tfsdk:"password"`
+	RetryWaitMinSec   types.Int64  `tfsdk:"retry_wait_min_sec"`
+	RetryWaitMaxSec   types.Int64  `tfsdk:"retry_wait_max_sec"`
+	RetryMax          types.Int64  `tfsdk:"retry_max"`
+	RequestTimeoutSec types.Int64  `tfsdk:"request_timeout_sec"`
+}
+
+type upcloudProvider struct{}
+
+var _ provider.Provider = New()
+
+func New() provider.Provider {
+	return &upcloudProvider{}
+}
+
+func (p *upcloudProvider) Metadata(_ context.Context, _ provider.MetadataRequest, resp *provider.MetadataResponse) {
+	resp.TypeName = "upcloud"
+	resp.Version = config.Version
+}
+
+func (p *upcloudProvider) Schema(_ context.Context, _ provider.SchemaRequest, resp *provider.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		Attributes: map[string]schema.Attribute{
+			"username": schema.StringAttribute{
+				Description: usernameDescription,
 				Optional:    true,
-				DefaultFunc: schema.EnvDefaultFunc("UPCLOUD_USERNAME", nil),
-				Description: "UpCloud username with API access. Can also be configured using the `UPCLOUD_USERNAME` environment variable.",
 			},
-			"password": {
-				Type:        schema.TypeString,
+			"password": schema.StringAttribute{
+				Description: passwordDescription,
 				Optional:    true,
-				DefaultFunc: schema.EnvDefaultFunc("UPCLOUD_PASSWORD", nil),
-				Description: "Password for UpCloud API user. Can also be configured using the `UPCLOUD_PASSWORD` environment variable.",
 			},
-			"retry_wait_min_sec": {
-				Type:        schema.TypeInt,
+			"retry_wait_min_sec": schema.Int64Attribute{
 				Optional:    true,
-				Default:     1,
 				Description: "Minimum time to wait between retries",
 			},
-			"retry_wait_max_sec": {
-				Type:        schema.TypeInt,
+			"retry_wait_max_sec": schema.Int64Attribute{
 				Optional:    true,
-				Default:     30,
 				Description: "Maximum time to wait between retries",
 			},
-			"retry_max": {
-				Type:        schema.TypeInt,
+			"retry_max": schema.Int64Attribute{
 				Optional:    true,
-				Default:     4,
 				Description: "Maximum number of retries",
 			},
-			"request_timeout_sec": {
-				Type:        schema.TypeInt,
+			"request_timeout_sec": schema.Int64Attribute{
 				Optional:    true,
-				Default:     120,
-				Description: "The duration (in seconds) that the provider waits for an HTTP request towards UpCloud API to complete. Defaults to 120 seconds",
+				Description: requestTimeoutDescription,
 			},
 		},
-
-		ResourcesMap: map[string]*schema.Resource{
-			"upcloud_server":                                  server.ResourceServer(),
-			"upcloud_server_group":                            servergroup.ResourceServerGroup(),
-			"upcloud_router":                                  router.ResourceRouter(),
-			"upcloud_storage":                                 storage.ResourceStorage(),
-			"upcloud_firewall_rules":                          firewall.ResourceFirewallRules(),
-			"upcloud_tag":                                     tag.ResourceTag(),
-			"upcloud_network":                                 network.ResourceNetwork(),
-			"upcloud_gateway":                                 gateway.ResourceGateway(),
-			"upcloud_gateway_connection":                      gateway.ResourceConnection(),
-			"upcloud_gateway_connection_tunnel":               gateway.ResourceTunnel(),
-			"upcloud_floating_ip_address":                     ip.ResourceFloatingIPAddress(),
-			"upcloud_object_storage":                          objectstorage.ResourceObjectStorage(),
-			"upcloud_managed_database_postgresql":             database.ResourcePostgreSQL(),
-			"upcloud_managed_database_mysql":                  database.ResourceMySQL(),
-			"upcloud_managed_database_redis":                  database.ResourceRedis(),
-			"upcloud_managed_database_opensearch":             database.ResourceOpenSearch(),
-			"upcloud_managed_database_user":                   database.ResourceUser(),
-			"upcloud_managed_database_logical_database":       database.ResourceLogicalDatabase(),
-			"upcloud_managed_object_storage":                  managedobjectstorage.ResourceManagedObjectStorage(),
-			"upcloud_managed_object_storage_policy":           managedobjectstorage.ResourceManagedObjectStoragePolicy(),
-			"upcloud_managed_object_storage_user":             managedobjectstorage.ResourceManagedObjectStorageUser(),
-			"upcloud_managed_object_storage_user_access_key":  managedobjectstorage.ResourceManagedObjectStorageUserAccessKey(),
-			"upcloud_managed_object_storage_user_policy":      managedobjectstorage.ResourceManagedObjectStorageUserPolicy(),
-			"upcloud_loadbalancer":                            loadbalancer.ResourceLoadBalancer(),
-			"upcloud_loadbalancer_resolver":                   loadbalancer.ResourceResolver(),
-			"upcloud_loadbalancer_backend":                    loadbalancer.ResourceBackend(),
-			"upcloud_loadbalancer_backend_tls_config":         loadbalancer.ResourceBackendTLSConfig(),
-			"upcloud_loadbalancer_static_backend_member":      loadbalancer.ResourceStaticBackendMember(),
-			"upcloud_loadbalancer_dynamic_backend_member":     loadbalancer.ResourceDynamicBackendMember(),
-			"upcloud_loadbalancer_frontend":                   loadbalancer.ResourceFrontend(),
-			"upcloud_loadbalancer_frontend_rule":              loadbalancer.ResourceFrontendRule(),
-			"upcloud_loadbalancer_frontend_tls_config":        loadbalancer.ResourceFrontendTLSConfig(),
-			"upcloud_loadbalancer_dynamic_certificate_bundle": loadbalancer.ResourceDynamicCertificateBundle(),
-			"upcloud_loadbalancer_manual_certificate_bundle":  loadbalancer.ResourceManualCertificateBundle(),
-			"upcloud_kubernetes_cluster":                      kubernetes.ResourceCluster(),
-			"upcloud_kubernetes_node_group":                   kubernetes.ResourceNodeGroup(),
-		},
-
-		DataSourcesMap: map[string]*schema.Resource{
-			"upcloud_zone":               cloud.DataSourceZone(),
-			"upcloud_zones":              cloud.DataSourceZones(),
-			"upcloud_networks":           network.DataSourceNetworks(),
-			"upcloud_hosts":              cloud.DataSourceHosts(),
-			"upcloud_ip_addresses":       ip.DataSourceIPAddresses(),
-			"upcloud_tags":               tag.DataSourceTags(),
-			"upcloud_storage":            storage.DataSourceStorage(),
-			"upcloud_kubernetes_cluster": kubernetes.DataSourceCluster(),
-			"upcloud_managed_database_opensearch_indices":  database.DataSourceOpenSearchIndices(),
-			"upcloud_managed_database_mysql_sessions":      database.DataSourceSessionsMySQL(),
-			"upcloud_managed_database_postgresql_sessions": database.DataSourceSessionsPostgreSQL(),
-			"upcloud_managed_database_redis_sessions":      database.DataSourceSessionsRedis(),
-			"upcloud_managed_object_storage_regions":       managedobjectstorage.DataSourceManagedObjectStorageRegions(),
-			"upcloud_managed_object_storage_policies":      managedobjectstorage.DataSourceManagedObjectStoragePolicies(),
-		},
-
-		ConfigureContextFunc: providerConfigure,
 	}
 }
 
-func providerConfigure(_ context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {
-	var diags diag.Diagnostics
+func withDefault[T comparable](val T, def T) T {
+	var empty T
+	if val == empty {
+		return def
+	}
+	return val
+}
 
-	requestTimeout := time.Duration(d.Get("request_timeout_sec").(int)) * time.Second
+func withEnvDefault(val string, env string) string {
+	return withDefault(val, os.Getenv(env))
+}
 
+func (p *upcloudProvider) Configure(ctx context.Context, req provider.ConfigureRequest, resp *provider.ConfigureResponse) {
+	var model upcloudProviderModel
+	if diags := req.Config.Get(ctx, &model); diags.HasError() {
+		resp.Diagnostics = append(resp.Diagnostics, diags...)
+		return
+	}
+
+	requestTimeout := time.Duration(withDefault(model.RequestTimeoutSec.ValueInt64(), 120)) * time.Second
 	config := Config{
-		Username: d.Get("username").(string),
-		Password: d.Get("password").(string),
+		Username: withEnvDefault(model.Username.ValueString(), "UPCLOUD_USERNAME"),
+		Password: withEnvDefault(model.Password.ValueString(), "UPCLOUD_PASSWORD"),
 	}
 
 	httpClient := retryablehttp.NewClient()
-	httpClient.RetryWaitMin = time.Duration(d.Get("retry_wait_min_sec").(int)) * time.Second
-	httpClient.RetryWaitMax = time.Duration(d.Get("retry_wait_max_sec").(int)) * time.Second
-	httpClient.RetryMax = d.Get("retry_max").(int)
+	httpClient.RetryWaitMin = time.Duration(withDefault(model.RetryWaitMinSec.ValueInt64(), 1)) * time.Second
+	httpClient.RetryWaitMax = time.Duration(withDefault(model.RetryWaitMaxSec.ValueInt64(), 30)) * time.Second
+	httpClient.RetryMax = int(withDefault(model.RetryMax.ValueInt64(), 4))
 
 	service := newUpCloudServiceConnection(
-		d.Get("username").(string),
-		d.Get("password").(string),
+		config.Username,
+		config.Password,
 		httpClient.HTTPClient,
 		requestTimeout,
 	)
 
 	_, err := config.checkLogin(service)
 	if err != nil {
-		return nil, diag.FromErr(err)
+		resp.Diagnostics.AddError("Authentication failed", "Failed to authenticate to UpCloud API with given credentials")
 	}
-
-	return service, diags
 }
 
-func newUpCloudServiceConnection(username, password string, httpClient *http.Client, requestTimeout time.Duration) *service.Service {
-	providerClient := client.New(
-		username,
-		password,
-		client.WithHTTPClient(httpClient),
-		client.WithTimeout(requestTimeout),
-	)
+func (p *upcloudProvider) Resources(_ context.Context) []func() resource.Resource {
+	return []func() resource.Resource{}
+}
 
-	providerClient.UserAgent = fmt.Sprintf("terraform-provider-upcloud/%s", config.Version)
-
-	return service.New(providerClient)
+func (p *upcloudProvider) DataSources(_ context.Context) []func() datasource.DataSource {
+	return []func() datasource.DataSource{}
 }
